feat(swap-cli-changelly): copilot comments
